#0: Created new test file for conv related slice tests
#0: Updated CMake for padded_slice and slice_write
